refactor(meal_planning): move command handlers to commands module
Move `replace_meal` and `regenerate_meal_plan` handler functions from
lib.rs to commands.rs for better code organization and consistency with
other domain crates (user, recipe).

Changes:
- Move replace_meal() from lib.rs to commands.rs
- Move regenerate_meal_plan() from lib.rs to commands.rs
- Add necessary imports to commands.rs (MealPlanAggregate, algorithm types, etc.)
- Update lib.rs pub use to re-export handler functions from commands module
- Maintain backward compatibility - functions still accessible via meal_planning::*

Benefits:
- Consistent module structure across all domain crates
- Command structs and handlers colocated in commands.rs
- All commands use evento::load for consistent data from event store
- Tests continue to work without modification

Note: ArchiveMealPlanCommand handler not yet implemented (planned feature)
